refactor(Testcase): service 타입 은닉 추가
service 타입 은닉을 추가함
